[#32] feat: AlarmSetupViewModel 구현
